refactor: gemini리뷰 반영, schedule과 estimatedCompletion 포멧 로직 통일(#108)
